made new component for badge
made new component for badge
