+++ app/Components/Badge/Badge.module.scss
@@ -1,12 +1,9 @@
-.badge {
-  display: flex;
-  width: 100%;
-}
+
 
 .badgeContent {
   display: flex;
   gap: 8px;
-  width: 100%;
+  width: 1000px;
 }
 
 .badgeContent span {
