+++ app/page.tsx
@@ -3,28 +3,14 @@
 import React from "react";
 import AsideMenu from "./Components/Aside/Aside";
 import styles from "./page.module.scss";
-import Navigation from "./Components/Navigation/Nav";
-import MusicListItem from "./Components/MusicList/MusicListItem";
 import Badge from "./Components/Badge/Badge";
+import style from "../app/Components/Badge/Badge.module.scss";
 
 const Home = () => {
-  const badgeItems = ["All", "Album", "Playlists", "Artists", "Downloaded"];
-  const [activeBadge, setActiveBadge] = useState("All");
 
-  const activeIndex = badgeItems.indexOf(activeBadge);
-
-  const badgeElements = badgeItems.map((item, index) => (
-    <span
-      key={index}
-      className={index === activeIndex ? style.active : ""}
-      onClick={() => setActiveBadge(item)}
-    >
-      {item}
-    </span>
-  ));
   return (
     <div className={styles.mainContent}>
-      <Badge badgeItems={badgeItems} activeIndex={activeIndex} />
+      <Badge>{badgeElements}</Badge>
     </div>
   );
 };
